+++ ribs/_utils.py
@@ -25,6 +25,21 @@ def parse_float_dtype(dtype):
     raise ValueError("Unsupported dtype. Must be np.float32 or np.float64")
 
 
+def np_scalar(scalar, dtype):
+    """Casts the scalar to the given numpy dtype.
+
+    This is useful for making sure all our scalars are in the correct dtype.
+
+    It is possible to just use np.array(scalar, dtype=dtype) since
+    zero-dimensional arrays behave like scalars, but the scalar would still show
+    up as an array if calling type().
+
+    It is also possible to use np.array(scalar, dtype=dtype).item(), but item()
+    converts outputs to standard Python objects, not numpy scalars.
+    """
+    return np.array([scalar], dtype=dtype)[0]
+
+
 def check_finite(x, name):
     """Checks that x is finite (i.e. not infinity or NaN).
 
@@ -203,7 +218,8 @@ def validate_single(archive, data):
     check_shape(data["solution"], "solution", archive.solution_dim,
                 "solution_dim")
 
-    data["objective"] = archive.dtype(data["objective"])
+    data["objective"] = np_scalar(data["objective"],
+                                  archive.dtypes["objective"])
     check_finite(data["objective"], "objective")
 
     data["measures"] = np.asarray(data["measures"])