+++ HISTORY.md
@@ -4,6 +4,11 @@
 
 ### Changelog
 
+#### API
+
+- Replace archive.dtype with archive.dtypes dict that holds dtype of every field
+  ({pr}`470`)
+
 #### Improvements
 
 - Upgrade setup-miniconda to v3 due to deprecation ({pr}`464`)